functional refactor!!
